Backend/requirements.txt
Django>=4.2
djangorestframework>=3.16.1
psycopg2-binary>=2.9
python-dotenv>=1.0,<2.0
coreapi>=2.3,<3.0
django-cors-headers>=4.0,<5.0
python-decouple>=3.8
